VcfFile -> VcfFileCheck  VcfFileHandler -> VcfFile
