+++ src/BedpeFilter/main.cpp
@@ -2,7 +2,7 @@
 #include "ToolBase.h"
 #include "NGSHelper.h"
 #include "Settings.h"
-#include "VcfFile.h"
+#include "VcfFileCheck.h"
 #include "BedpeFile.h"
 #include <QTextStream>
 #include <QFileInfo>
