+++ src/VcfFilter/main.cpp
@@ -1,7 +1,7 @@
 #include "Exceptions.h"
 #include "ToolBase.h"
 #include "BedFile.h"
-#include "VcfFile.h"
+#include "VcfFileCheck.h"
 #include "ChromosomalIndex.h"
 #include "Helper.h"
 #include <QFile>
@@ -268,9 +268,9 @@ public:
             ///Filter by region. First return CHROM and POS and then remove all lines that do not satisfy the check
             if (reg != "")
             {
-				QByteArray chr = VcfFile::getPartByColumn(line, VcfFile::CHROM);
-				QByteArray start = VcfFile::getPartByColumn(line, VcfFile::POS);
-				QByteArray ref = VcfFile::getPartByColumn(line, VcfFile::REF);
+				QByteArray chr = VcfFileCheck::getPartByColumn(line, VcfFileCheck::CHROM);
+				QByteArray start = VcfFileCheck::getPartByColumn(line, VcfFileCheck::POS);
+				QByteArray ref = VcfFileCheck::getPartByColumn(line, VcfFileCheck::REF);
 
 				if (roi_index.matchingIndex(chr, start.toInt(), start.toInt() + ref.length())==-1)
                 {
@@ -281,8 +281,8 @@ public:
             ///Filter by variant_type.
             if (variant_type != "")
 			{
-				QByteArray ref = VcfFile::getPartByColumn(line, VcfFile::REF).trimmed();
-				QByteArray alt = VcfFile::getPartByColumn(line, VcfFile::ALT).trimmed();
+				QByteArray ref = VcfFileCheck::getPartByColumn(line, VcfFileCheck::REF).trimmed();
+				QByteArray alt = VcfFileCheck::getPartByColumn(line, VcfFileCheck::ALT).trimmed();
 
                 QString type;
                 if (ref.length() == 1 && alt.length() == 1)
@@ -316,7 +316,7 @@ public:
             ///Filter by QUALITY. Return QUAL and then remove all lines that do not satisfy the check
             if (quality != 0.0)
             {
-				QByteArray qual = VcfFile::getPartByColumn(line, VcfFile::QUAL).trimmed();
+				QByteArray qual = VcfFileCheck::getPartByColumn(line, VcfFileCheck::QUAL).trimmed();
 				bool ok;
 				double qual_value = qual.toDouble(&ok);
 				if (!ok && qual!=".") THROW(ProgrammingException, "Quality '" + qual + "' cannot be converted to a number in line: " + line);
@@ -329,7 +329,7 @@ public:
             ///Filter by empty filters (will remove empty filters).
             if (filter_empty)
             {
-				QByteArray filter = VcfFile::getPartByColumn(line, VcfFile::FILTER).trimmed();
+				QByteArray filter = VcfFileCheck::getPartByColumn(line, VcfFileCheck::FILTER).trimmed();
 
 				if (filter!="." && filter!="" && filter!="PASS")
                 {
@@ -340,7 +340,7 @@ public:
             ///Filter FILTER column via regex
 			if (filter != "")
             {
-				QByteArray filter = VcfFile::getPartByColumn(line, VcfFile::FILTER).trimmed();
+				QByteArray filter = VcfFileCheck::getPartByColumn(line, VcfFileCheck::FILTER).trimmed();
                 auto match = filter_re.match(filter);
                 if (!match.hasMatch())
                 {
@@ -351,7 +351,7 @@ public:
             ///Filter ID column via regex
 			if (id != "")
             {
-				QByteArray id = VcfFile::getPartByColumn(line, VcfFile::ID).trimmed();
+				QByteArray id = VcfFileCheck::getPartByColumn(line, VcfFileCheck::ID).trimmed();
                 auto match = id_re.match(id);
                 if (!match.hasMatch())
                 {
@@ -362,7 +362,7 @@ public:
 			///Filter by info operators in INFO column
 			if (!info_filters.isEmpty())
             {
-				QByteArrayList info_parts = VcfFile::getPartByColumn(line, VcfFile::INFO).trimmed().split(';');
+				QByteArrayList info_parts = VcfFileCheck::getPartByColumn(line, VcfFileCheck::INFO).trimmed().split(';');
 
 				bool passes_filters = true;
 				foreach(const QByteArray& info_part, info_parts)
@@ -393,13 +393,13 @@ public:
             ///Filter by sample operators in the SAMPLE column
 			if (!sample_filters.isEmpty())
             {
-				QByteArrayList format_entries = VcfFile::getPartByColumn(line, VcfFile::FORMAT).trimmed().split(':');
+				QByteArrayList format_entries = VcfFileCheck::getPartByColumn(line, VcfFileCheck::FORMAT).trimmed().split(':');
 
 				int samples_passing = 0;
 				int samples_failing = 0;
-				for (int i = VcfFile::MIN_COLS + 1; i < column_count; ++i)
+				for (int i = VcfFileCheck::MIN_COLS + 1; i < column_count; ++i)
 				{
-					QByteArrayList sample_parts = VcfFile::getPartByColumn(line, i).split(':');
+					QByteArrayList sample_parts = VcfFileCheck::getPartByColumn(line, i).split(':');
 
 					bool current_sample_passes = true;
 					foreach(const FilterDefinition& filter, sample_filters)
